requirements.txt
pydantic==2.10.6
numpy==1.26.4
torch==2.2.0
transformers==4.47.0
layoutparser[effdet]
opencv-python==4.5.5.64
opencv-python-headless==4.5.4.60
Pillow==9.5.0
ipython
openai
albumentations==1.1.0
matplotlib>=3.5.3
SmilesPE==0.0.3
OpenNMT-py==2.2.0
rdkit-pypi>=2021.03.2
timm==0.4.12
pandas>=1.2.4
boto3>=1.34
pyarrow>=15.0.0
pycocotools>=2.0.4
pytorch-lightning>=1.8.6
huggingface-hub>=0.11.0
hf_xet
easyocr>=1.6.2
cairosvg
PyPDF2
pdfminer.six==20221105
pdf2image
pytesseract
safetensors
einops
chemrxnextractor @ git+https://github.com/CrystalEye42/ChemRxnExtractor.git@0f9529dbe3656e4ef5ea96c5b5ba990f7481700b
anthropic

gradio>=4.44.0
